refactor: move memoize to separate file
