+++ tsconfig.json
@@ -52,6 +52,7 @@
     "src/fs/unsymlink.ts",
     "src/getSaveType.ts",
     "src/index.ts",
+    "src/memoize.ts",
     "src/install/fetch.ts",
     "src/install/flattenDependencies.ts",
     "src/install/installMultiple.ts",
